Air Movement and Enemy Attacking
Modified how player movement works in the air to prevent the player from moving as fast when they're in the air. Also allowed enemies to hit the player, knocking the player backwards. The player does not take damage yet. Also started work on modifying the battlezones to, instead of spawning all the enemies at once, spawn them in waves.